less side-effects, add analyzers and formatters
Ok so this patch was mainly concerned with having a generic (and
publicly-exposed) approach to decorating PropTypes with introspection.

It is definitely an API-breaking patch; the properties "shape", "oneOf",
and "oneOfType" that were previously stamped on their corresponding
PropType checkers no longer exist; instead we have a generic $meta
object that contains these context variables in the "args" property.

Changes:

- the library can now handle introspecting & formatting of custom
  PropTypes

- added introspection support for `arrayOf`

- implicit definition of custom toString() was dropped in favor of
  explicit formatting; one can now choose to use the formatter directly
  if they need to stringify a prop (and implement their own formatters
  if we do not support them out of the box)

- added "analyzers" that give us something close to an AST of the
  propType nodes (this is opt-in for those who need it just like the
  formatters)

- added more test coverage

- simplified the validator code (got rid of the class, a single function
  for validating is easier to reason about for users like myself)